Fix: User -> Member 네이밍 통일 및 역할분리
