+++ src/main/java/org/zapply/product/domain/user/controller/AuthController.java
@@ -15,7 +15,7 @@ import org.zapply.product.domain.user.dto.response.MemberResponse;
 import org.zapply.product.domain.user.dto.response.TokenResponse;
 import org.zapply.product.domain.user.service.AccountService;
 import org.zapply.product.domain.user.service.AuthService;
-import org.zapply.product.domain.user.service.UserService;
+import org.zapply.product.domain.user.service.MemberService;
 import org.zapply.product.global.apiPayload.response.ApiResponse;
 import org.zapply.product.global.security.AuthDetails;
 import org.zapply.product.global.security.jwt.JwtProvider;
@@ -26,13 +26,11 @@ import org.zapply.product.global.security.jwt.JwtProvider;
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final UserService userService;
     @Value("${jwt.header}")
     private String tokenHeader;
 
     private final AuthService authService;
     private final JwtProvider jwtProvider;
-    private final AccountService accountService;
 
     @PostMapping("/sign-up")
     @Operation(summary = "회원가입", description = "사용자의 정보 입력 후 회원가입")
@@ -67,25 +65,6 @@ public class AuthController {
     @Operation(summary = "이메일 중복 확인", description = "이메일 중복 확인. true - 중복, false - 사용 가능")
     public ApiResponse<Boolean> checkEmailDuplicate(@RequestParam("email") String email) {
         return ApiResponse.success(authService.checkEmailDuplicate(email));
-    }
-
-    @GetMapping()
-    @Operation(summary = "사용자 정보 조회(이름, 이메일)", description = "jwt 토큰을 통해 사용자 정보 조회")
-    public ApiResponse<MemberResponse> getUserInformation(@AuthenticationPrincipal AuthDetails authDetails) {
-        return ApiResponse.success(MemberResponse.of(authDetails.getMember()));
-    }
-
-    // 사용자가 연동한 account를 조회
-    @GetMapping("/accounts")
-    @Operation(summary = "사용자 연동 계정 조회", description = "사용자가 연동한 계정 조회")
-    public ApiResponse<?> getUserAccount(@AuthenticationPrincipal AuthDetails authDetails) {
-        return ApiResponse.success(accountService.getAccountsInfo(authDetails.getMember()));
-    }
-
-    @PatchMapping("/mypage")
-    @Operation(summary = "사용자 이름 수정", description = "사용자 이름 수정")
-    public ApiResponse<?> updateUserName(@AuthenticationPrincipal AuthDetails authDetails, @RequestParam("name") String name) {
-        return ApiResponse.success(userService.updateUserName(authDetails.getMember(), name));
     }
 }
