+++ src/main/java/org/zapply/product/domain/user/service/AuthService.java
@@ -22,7 +22,7 @@ import org.zapply.product.global.security.jwt.JwtProvider;
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class AuthService {
-    private final UserService userService;
+    private final MemberService memberService;
     private final CredentialService credentialService;
     private final JwtProvider jwtProvider;
     private final RedisClient redisClient;
@@ -40,7 +40,7 @@ public class AuthService {
     @Transactional
     public MemberResponse signUp(AuthRequest authRequest) {
         Credential credential = credentialService.createCredential(authRequest);
-        Member member = userService.createUser(credential, authRequest);
+        Member member = memberService.createMember(credential, authRequest);
 
         return MemberResponse.of(member);
     }
@@ -52,7 +52,7 @@ public class AuthService {
      */
     @Transactional
     public LoginResponse signIn(SignInRequest signInRequest) {
-        Member member = userService.getUserByEmail(signInRequest.email());
+        Member member = memberService.getMemberByEmail(signInRequest.email());
         credentialService.checkPassword(member, signInRequest.password());
         TokenResponse tokenResponse = jwtProvider.createToken(member);
         redisClient.setValue(member.getEmail(), tokenResponse.refreshToken(), refreshTokenExpirationTime);
